touchup review service

+++ backend/webshop/review/src/main/java/com/fer/hr/review/model/AverageRatingAndCount.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class AverageRatingAndCount {
-    private Float averageRating;
+    private double averageRating;
     private int count;
 }
